+++ asyncdex/models/group.py
@@ -3,6 +3,7 @@ from typing import Any, Dict, TYPE_CHECKING
 from .abc import GenericModelList, Model
 from .mixins import DatetimeMixin
 from .user import User
+from ..constants import routes
 from ..utils import copy_key_to_attribute
 
 if TYPE_CHECKING:
@@ -56,4 +57,44 @@ class Group(Model, DatetimeMixin):
         await self.client.batch_chapters(*self.chapters)
 
     async def fetch(self):
+        """Fetch data about the group. |permission| ``group.view``
+
+        :raises: :class:`.InvalidID` if a group with the ID does not exist.
+        """
         await self._fetch("scanlation_group.view", "group")
+
+    async def update(self):
+        """Update the scanlation group. |auth|
+
+        .. versionadded:: 0.5
+        """
+        params = {"name": self.name, "members": [item.id for item in self.members], "leader": self.leader.id}
+        self.client.raise_exception_if_not_authenticated("PUT", routes["group"])
+        r = await self.client.request("PUT", routes["group"].format(id=self.id), json=params)
+        json = await r.json()
+        r.close()
+        obj = type(self)(self.client, data=json)
+        self.transfer(obj)
+
+    async def delete(self):
+        """Delete the scanlation group. |auth|
+
+        .. versionadded:: 0.5
+        """
+        return await self._delete("group")
+
+    async def follow(self):
+        """Follow the scanlation group. |auth|
+
+        .. versionadded:: 0.5
+        """
+        self.client.raise_exception_if_not_authenticated("POST", routes["group_follow"])
+        (await self.client.request("POST", routes["group_follow"].format(id=self.id))).close()
+
+    async def unfollow(self):
+        """Unfollow the scanlation group. |auth|
+
+        .. versionadded:: 0.5
+        """
+        self.client.raise_exception_if_not_authenticated("DELETE", routes["group_follow"])
+        (await self.client.request("DELETE", routes["group_follow"].format(id=self.id))).close()